# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  helpers.Response:
    properties:
      data: {}
      error: {}
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  users.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
  users.ForgotUsernameRequest:
    properties:
      email:
        type: string
    type: object
  users.RequestChangePassword:
    properties:
      current_password:
        type: string
      new_password:
        type: string
      new_password_confirm:
        type: string
    type: object
  users.RequestCreateUser:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  users.RequestUpdateUser:
    properties:
      email:
        type: string
      mobile_phone:
        type: string
      status:
        type: integer
      username:
        type: string
    type: object
  users.RequestUserRoles:
    properties:
      role_ids:
        items:
          type: integer
        type: array
    type: object
  users.ResetPasswordRequest:
    properties:
      password:
        type: string
      password_confirmation:
        type: string
      token:
        type: string
    type: object
...
info:
  contact: {}
  description: Go API service
  title: Go API
  version: "1.0"
paths:
...
  /users/:
    get:
      parameters:
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 10)
        in: query
        name: limit
        type: integer
      - description: Order by
        in: query
        name: order_by
        type: string
      - description: Search keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Get user detail
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.RequestUpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/{id}/roles:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Get role IDs of user
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.RequestUserRoles'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Update roles of user
      tags:
      - users
...
  /users/change-password:
    put:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.RequestChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Change own password
      tags:
      - users
  /users/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.RequestCreateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - users
  /users/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Request password reset code by email
      tags:
      - users
  /users/forgot-username:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.ForgotUsernameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Send username to email
      tags:
      - users
  /users/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
  /users/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Reset password with token
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by the access token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"